+++ backend/sswm/src/main/java/com/ground/sswm/studyroom/dto/StudyroomDto.java
@@ -23,7 +23,7 @@ public class StudyroomDto {
     private int studyAvgTime;
     private String image;
     private boolean isDeleted;
-    private int createdAt;
+    private long createdAt;
     private List<TagDto> tags;
 
     public boolean getIsPublic() {
@@ -37,7 +37,7 @@ public class StudyroomDto {
     @Builder
     public StudyroomDto(Long id, String name, String notice, boolean isPublic, String enterCode,
         int maxUserNum,
-        int maxRestTime, int studyAvgTime, String image, boolean isDeleted, int createdAt) {
+        int maxRestTime, int studyAvgTime, String image, boolean isDeleted, long createdAt) {
         this.id = id;
         this.name = name;
         this.notice = notice;
